+++ lib/components/dataset/DatasetRouter.js
@@ -4,7 +4,7 @@ import { withRouter } from 'react-router-dom'
 
 import PropTypes from 'prop-types'
 
-import DatasetRefProps from '../propTypes/datasetRefProps'
+import DatasetRefProps from '../../propTypes/datasetRefProps'
 
 import Base from '../Base'
