+++ lib/components/dataset/Overview.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types'
 import MetaProps from '../../propTypes/metaProps.js'
 import Base from '../Base'
 import TagList from '../TagList'
-import StatsLineLarge from './StatsLineLarge'
+import StatsLineLarge from '../StatsLineLarge'
 import CommitItem from '../item/CommitItem'
 import Button from '../Button'
