+++ lib/components/stylesheet/Swatch.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import Base from './Base'
+import Base from '../Base'
 
 export default class Swatch extends Base {
   template (css) {
