+++ lib/containers/Transform.js
@@ -2,7 +2,7 @@ import { connect } from 'react-redux'
 
 import { selectCAFSString, selectCAFSLoading, selectCAFSError } from '../selectors/cafs'
 import { loadTransform } from '../actions/dataset'
-import Transform from '../components/transform'
+import Transform from '../components/dataset/transform'
 
 const TransformContainer = connect(
   (state, ownProps) => {
