+++ lib/containers/Viz.js
@@ -2,7 +2,7 @@ import { connect } from 'react-redux'
 
 import { selectCAFSString, selectCAFSLoading, selectCAFSError } from '../selectors/cafs'
 import { loadViz } from '../actions/dataset'
-import Viz from '../components/viz'
+import Viz from '../components/dataset/viz'
 
 const VizContainer = connect(
   (state, ownProps) => {
