+++ lib/routes.js
@@ -8,7 +8,7 @@ import DatasetContainer from './containers/Dataset'
 import AddDatasetContainer from './containers/AddDataset'
 import ProfilesContainer from './containers/Profiles'
 import ProfileContainer from './containers/Profile'
-import Stylesheet from './components/Stylesheet'
+import Stylesheet from './components/stylesheet/Stylesheet'
 // import SearchResultsContainer from './containers/SearchResults'
 import MetadataEditorContainer from './containers/MetadataEditor'
 import EditorContainer from './containers/Editor'
